refactor(tui::ui::components::popups::youtube_search): move function "*mount_youtube_search*" to popup file
